Remove `unveil` JavaScript library
The `unveil` library is used to lazy-load images, and its last release
was in 2013. This is useful functionality, but we don't need JavaScript
to get it. All modern browsers now support it natively with the
`loading="lazy"` attribute on the `img` HTML tag:

- https://caniuse.com/loading-lazy-attr
- https://developer.mozilla.org/en-US/docs/Web/HTML/Element/img#loading

This patch removes the `unveil` library and all references to it and
uses the `loading` attribute in the one place where `unveil` was being
used. This includes `fundraising-heroes.js`, the sole purpose of which
was to initialize `unveil`.

It seems that `unveil` was also trying to swap out higher-quality images
on retina display. My sense is that this is not needed, and the
associated attributes have been removed.